refactor(core): extract service test fixtures into test_support module
Downstream crates (Tauri, CLI) need to construct MarketplaceService
instances and seed marketplace registries to unit-test their command
handlers. The fixtures that do this (temp_service, make_plugin_with_skills,
relative_path_entry, seed_marketplace_with_registry, PanicOnNetworkBackend)
had been confined to service::browse::tests — reachable only from inside
the crate.

Moves them into a feature-gated pub module at service::test_support,
parallel to the existing crate-level test_utils module. The test-support
feature was already declared; this extends its scope. Downstream crates
opt in via a dev-dependencies override on kiro-market-core with the
test-support feature enabled, keeping the fixtures out of production
builds.

tempfile moves from dev-deps-only to an optional regular dependency
activated by the test-support feature, so the fixtures compile when
the feature is enabled from outside the crate.

Part of #43.
