+++ crates/kiro-market-core/Cargo.toml
@@ -12,7 +12,7 @@ workspace = true
 default = []
 cli = ["clap"]
 specta = ["dep:specta"]
-test-support = []
+test-support = ["dep:tempfile"]
 
 [dependencies]
 serde = { workspace = true }
@@ -36,6 +36,7 @@ specta = { version = "2.0.0-rc.24", optional = true, features = ["derive", "serd
 # the resolved feature list contains `ssl` so this regression cannot
 # return unnoticed.
 curl = { workspace = true }
+tempfile = { workspace = true, optional = true }
 
 [target.'cfg(windows)'.dependencies]
 junction = { workspace = true }
